--- frontend/src/app/admin/news/hooks/index.js
@@ -1,4 +1,0 @@
-export { useNewsOperations } from './useNewsOperations';
-export { useNewsFilters } from './useNewsFilters';
-export { useNewsModals } from './useNewsModals';
-export { useNewsURL } from './useNewsURL';
