+++ frontend/src/app/admin/news/page.js
@@ -4,7 +4,10 @@ import { useState, useEffect, useCallback } from 'react';
 import { useSelector } from 'react-redux';
 import { selectCurrentAdmin } from '@/rtk/superadmin/adminSlice';
 import { useAdminNews } from '@/hooks/admin/useAdminData';
-import { useNewsOperations, useNewsFilters, useNewsModals, useNewsURL } from './hooks';
+import { useNewsOperations } from '@/hooks/admin/useNewsOperations';
+import { useNewsFilters } from '@/hooks/admin/useNewsFilters';
+import { useNewsModals } from '@/hooks/admin/useNewsModals';
+import { useNewsURL } from '@/hooks/admin/useNewsURL';
 import { NewsTable, CreatePostForm, SearchAndFilterControls, ViewDetailsModal } from './components';
 import { ErrorBoundary, SuccessModal } from '@/components';
 import { SkeletonLoader } from '../components';
