--- frontend/src/app/admin/organization/hooks/index.js
@@ -1,3 +1,0 @@
-export { useModalScrollLock } from './useModalScrollLock';
-export { useModalMessage } from './useModalMessage';
-export { useFormChanges } from './useFormChanges';
