+++ frontend/src/app/admin/programs/archive/page.js
@@ -11,7 +11,8 @@ import { SkeletonLoader } from '../../components';
 import { ConfirmationModal, ErrorBoundary, SuccessModal } from '@/components';
 import { handleApiError } from '@/utils/admin/errorHandler';
 import { TIMEOUTS } from '@/utils/admin/constants';
-import { useProgramsManagement, useModalManagement } from '../hooks';
+import { useProgramsManagement } from '@/hooks/admin/useProgramsManagement';
+import { useModalManagement } from '@/hooks/admin/useModalManagement';
 import { filterAndSortPrograms } from '@/utils/admin/programUtils';
 import styles from '../programs.module.css';
 import { FiArrowLeft } from 'react-icons/fi';
