--- frontend/src/app/admin/programs/hooks/index.js
@@ -1,9 +1,0 @@
-export { useProgramForm } from './useProgramForm';
-export { useImageUpload } from './useImageUpload';
-export { useCollaboration } from './useCollaboration';
-export { useCollaborationRequests } from './useCollaborationRequests';
-export { useProgramsManagement } from './useProgramsManagement';
-export { useProgramFilters } from './useProgramFilters';
-export { useModalManagement } from './useModalManagement';
-export { useCollaborationManagement } from './useCollaborationManagement';
-export { useProgramActions } from './useProgramActions';
