+++ frontend/src/app/admin/programs/page.js
@@ -5,14 +5,17 @@ import { useSelector } from 'react-redux';
 import { useSearchParams, useRouter } from 'next/navigation';
 import { selectCurrentAdmin } from '@/rtk/superadmin/adminSlice';
 import { useAdminPrograms } from '@/hooks/admin/useAdminData';
-import { useCollaborationRequests } from './hooks';
+import { useCollaborationRequests } from '@/hooks/admin/useCollaborationRequests';
 import { ViewDetailsModal, ProgramsContainer, CollaborationsContainer, SearchAndFilterControls } from './components';
 import ProgramForm from './components/ProgramForm/ProgramForm';
 import { SkeletonLoader } from '../components';
 import { ConfirmationModal, ErrorBoundary, SuccessModal } from '@/components';
 import { handleApiError } from '@/utils/admin/errorHandler';
 import { TIMEOUTS } from '@/utils/admin/constants';
-import { useProgramsManagement, useProgramFilters, useModalManagement, useCollaborationManagement } from './hooks';
+import { useProgramsManagement } from '@/hooks/admin/useProgramsManagement';
+import { useProgramFilters } from '@/hooks/admin/useProgramFilters';
+import { useModalManagement } from '@/hooks/admin/useModalManagement';
+import { useCollaborationManagement } from '@/hooks/admin/useCollaborationManagement';
 import styles from './programs.module.css';
 import { FaPlus, FaUsers } from 'react-icons/fa';
 import { FiArchive } from 'react-icons/fi';
