+++ frontend/src/hooks/admin/useCollaborationManagement.js
@@ -104,3 +104,4 @@ export const useCollaborationManagement = (
     handleCollaborationAction
   };
 };
+
