+++ frontend/src/hooks/admin/useCollaborationRequests.js
@@ -130,3 +130,4 @@ export const useCollaborationRequests = () => {
     declineCollaboration
   };
 };
+
