+++ frontend/src/hooks/admin/useFormChanges.js
@@ -57,3 +57,4 @@ export const useFormChanges = () => {
     hasSectionChangesFromData
   };
 };
+
