+++ frontend/src/hooks/admin/useModalManagement.js
@@ -113,3 +113,4 @@ export const useModalManagement = () => {
     resetEditMode
   };
 };
+
