+++ frontend/src/hooks/admin/useModalMessage.js
@@ -35,3 +35,4 @@ export const useModalMessage = (isOpen, autoClearDelay = 3000) => {
     clearMessage
   };
 };
+
