+++ frontend/src/hooks/admin/useModalScrollLock.js
@@ -34,3 +34,4 @@ export const useModalScrollLock = (isOpen) => {
     };
   }, [isOpen]);
 };
+
