+++ frontend/src/hooks/admin/useNewsOperations.js
@@ -622,3 +622,4 @@ export const useNewsOperations = (orgId, refreshNews, setSuccessModal) => {
     validateNewsData
   };
 };
+
