+++ frontend/src/hooks/admin/useNewsURL.js
@@ -90,3 +90,4 @@ export const useNewsURL = () => {
     updateURLParams
   };
 };
+
