+++ frontend/src/hooks/admin/useProgramFilters.js
@@ -134,3 +134,4 @@ export const useProgramFilters = (programs, collaborations) => {
     updateURLParams
   };
 };
+
