+++ frontend/src/hooks/admin/useProgramsManagement.js
@@ -857,3 +857,4 @@ export const useProgramsManagement = (currentAdmin, refreshPrograms, setSuccessM
     handleUnarchiveProgram
   };
 };
+
